use numba annotations njit

+++ y2024DES/data.py
@@ -12,7 +12,7 @@ selected_columns = data_frame[
 ]
 
 n = selected_columns["zHD"].size
-covariance_stat = covariance_file["cov_mu"].to_numpy().reshape((n, n))
+covariance_stat = covariance_file["cov_mu"].to_numpy(dtype=np.float64).reshape((n, n))
 
 # Add statistical uncertainties
 covariance_matrix = covariance_stat + np.diag(
@@ -20,8 +20,8 @@ covariance_matrix = covariance_stat + np.diag(
 )
 
 # Sort by redshift
-z_values = selected_columns["zHD"].to_numpy()
-z_hel_values = selected_columns["zHEL"].to_numpy()
+z_values = selected_columns["zHD"].to_numpy(dtype=np.float64)
+z_hel_values = selected_columns["zHEL"].to_numpy(dtype=np.float64)
 sort_indices = np.argsort(z_values)
 
 """
@@ -29,13 +29,16 @@ effective_sample_size:
 sum (1 - PROBCC_BEAMS) + low_z size => 1735
 """
 
-effective_sample_size = np.round((1 - selected_columns["PROBCC_BEAMS"]).sum()).astype(int)
+effective_sample_size = np.round((1 - selected_columns["PROBCC_BEAMS"]).sum()).astype(
+    int
+)
+
 
 def get_data():
     return (
         f"DES-SN5YR - effective: {effective_sample_size} SNe",
         z_values[sort_indices],
         z_hel_values[sort_indices],
-        selected_columns["MU"].to_numpy()[sort_indices],
+        selected_columns["MU"].to_numpy(dtype=np.float64)[sort_indices],
         covariance_matrix[sort_indices, :][:, sort_indices],
     )
